requirements.txt
openai
watchdog
python-dotenv
PyGithub
streamlit